renamed model/controller files, added TODO for common-password-checker
